+++ articles/flow/migration/_migration-banner.asciidoc
@@ -1,7 +1,15 @@
 [.toc-banner-source.skip-search-index]
 --
+image:images/migration-assistance.png[migration assistance]
+
+[.toc-banner-source-text]
+Are you planning on migrating your application? Get help from Vaadin experts.
+
+[.toc-banner-source-link]
+\https://vaadin.com/solutions/migration-assistance
+
 [source,html,role=render-only]
 ----
-include::{root}/frontend/demo/banners/migration/migration-banner.ts[render]
+include::{root}/frontend/demo/render-banner.ts[render]
 ----
 --
