+++ dspublisher/theme/styles/banners.css
@@ -2,8 +2,10 @@
   position: fixed;
   top: 0;
   left: 0;
+  width: 10px;
+  height: 10px;
   visibility: hidden;
-  height: 3px;
+  opacity: 0;
   z-index: -9999;
 }
